requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pyodbc==5.1.0
pydantic==2.9.0
python-dotenv==1.0.1
